+++ src/types.ts
@@ -562,8 +562,8 @@ export interface FeatureConfig {
   lyricsCacheTTL: number;
   /** Button to clear lyrics cache */
   clearLyricsCache: undefined;
-  /** Whether to use advanced filtering when searching for lyrics (exact, exact-ish) */
-  advancedLyricsFilter: boolean;
+  // /** Whether to use advanced filtering when searching for lyrics (exact, exact-ish) */
+  // advancedLyricsFilter: boolean;
 
   //#region misc
   /** The locale to use for translations */